rename btrfs2s3 run -> btrfs2s3 update
