+++ tests/commands/update/invocation_test.py::test_refuse_to_run_unattended_without_pretend_or_force.golden
