+++ tests/commands/update/print_actions_test.py::test_print_actions[3-3-3-3-3].golden
