+++ tests/commands/update/print_assessment_test.py::test_dont_keep.golden
