+++ tests/commands/update/print_assessment_test.py::test_most_recent.golden
