+++ tests/commands/update/print_assessment_test.py::test_new_proposed.golden
