+++ tests/commands/update/print_assessment_test.py::test_print_assessment.golden
