+++ tests/commands/update/print_assessment_test.py::test_send_ancestor.golden
